requirements.txt
--index-url https://pypi.org/simple/
numpy
opencv-python
PyAudio
pyzmq
Pillow
